refactor: 移动工具方法到独立的工具类
